requirements.txt
weaviate-client
scrapy